update ProblemConnection structure and AssignmentConnection structure
